+++ api/src/team/domain/usecases/accept-certification-center-invitation.usecase.js
@@ -1,4 +1,4 @@
-import { AlreadyExistingMembershipError } from '../../../src/shared/domain/errors.js';
+import { AlreadyExistingMembershipError } from '../../../shared/domain/errors.js';
 
 const acceptCertificationCenterInvitation = async function ({
   certificationCenterInvitationId,
